pgweb/requirements.txt
git+git://github.com/hpidcock/operator.git@master#egg=ops
ops-lib-pgsql
